move profile routes to own file
